Solving linear systems

Warning

The overall structure of Linear Solvers is expected to undergo a major update in the near future. In particular, a number of type parametrization shall eventually be replaced by a trait-based implementation, which should prove more flexible.

The high-level interface should not change too much, and the underlying linear algebra techniques shall remain the same.

Overview

The first system we consider is the symmetric quasi-definite augmented system

\[\begin{bmatrix} -(\Theta^{-1} + R_{p}) & A^{T}\\ A & R_{d} \end{bmatrix} \begin{bmatrix} \Delta x\\ \Delta y \end{bmatrix} = \begin{bmatrix} \xi_d\\ \xi_p \end{bmatrix}\]

One can pivot out the upper-left diagonal block to obtain the positive-definite normal equations system

\[\left( A (\Theta^{-1} + R_{p})^{-1} A^{T} + R_{d} \right) \Delta y = \xi_{p} + A (Θ^{-1} + R_{p})^{-1} \xi_{d}\]

Linear solvers

Here is a list of currently supported linear solvers:

Linear solver type	type of $A$	System	Method	Backend

DenseLinearSolver	Matrix	Normal Eqn	Direct (Cholesky)	LAPACK

SparseIndefLinearSolver	SparseMatricCSC	Augm. Sys	Direct (LDLt)	SuiteSparse

SparsePosDefLinearSolver	SparseMatricCSC	Normal Eqn	Direct (Cholesky)	SuiteSparse

LDLFLinearSolver	SparseMatricCSC	Augm. Sys	Direct (LDLt)	LDLFactorizations

AbstractLinearSolver

This is the base type from which all implementations should derive.

Tulip.TLPLinearAlgebra.AbstractLinearSolver — Type

AbstractLinearSolver{Tv, Ta}

Abstract container for linear solver used in solving the augmented system.

Custom linear solvers should inherit from the AbstractLinearSolver class, and extend the following two functions:

Tulip.TLPLinearAlgebra.update_linear_solver! — Function

update_linear_solver!(ls, θ, regP, regD)

Update internal data, and re-compute factorization/pre-conditioner.

After this call, ls can be used to solve the augmented system

    [-(Θ^{-1} + Rp)   A'] [dx] = [ξd]
    [   A             Rd] [dy] = [ξp]

for given right-hand sides ξd and ξp.

Tulip.TLPLinearAlgebra.solve_augmented_system! — Function

solve_augmented_system!(dx, dy, ls, ξp, ξd)

Solve the symmetric quasi-definite augmented system

    [-(Θ^{-1} + Rp)   A'] [dx] = [ξd]
    [   A             Rd] [dy] = [ξp]

and over-write dx, dy with the result.

Arguments

dx, dy: Vectors of unknowns, modified in-place

ls: Linear solver for the augmented system

ξp, ξd: Right-hand-side vectors

DenseLinearSolver

Tulip.TLPLinearAlgebra.DenseLinearSolver — Type

DenseLinearSolver{Tv}

Linear solver for the 2x2 augmented system

    [-(Θ^{-1} + Rp)   A'] [dx] = [xi_d]
    [   A             Rd] [dy] = [xi_p]

with $A$ dense.

The augmented system is automatically reduced to the normal equations system. BLAS/LAPACK functions are used whenever applicable.

Tulip.TLPLinearAlgebra.update_linear_solver! — Method

update_linear_solver!(ls::DenseLinearSolver{<:Real}, θ, regP, regD)

Compute normal equations system matrix and update Cholesky factorization.

Uses Julia's generic linear algebra.

Tulip.TLPLinearAlgebra.update_linear_solver! — Method

update_linear_solver!(ls::DenseLinearSolver{<:BlasReal}, θ, regP, regD)

Compute normal equations system matrix and update Cholesky factorization.

Uses BLAS and LAPACK routines.

Tulip.TLPLinearAlgebra.solve_augmented_system! — Method

solve_augmented_system!(dx, dy, ls, ξp, ξd)

Solve the augmented system, overwriting dx, dy with the result.

Uses two generic triangular solves for solving the normal equations system.

Tulip.TLPLinearAlgebra.solve_augmented_system! — Method

solve_augmented_system!(dx, dy, ls, ξp, ξd)

Solve the augmented system, overwriting dx, dy with the result.

Uses one LAPACK call for solving the normal equations system.

SparseIndefLinearSolver

Tulip.TLPLinearAlgebra.SparseIndefLinearSolver — Type

SparseIndefLinearSolver{Tv}

Linear solver for the 2x2 augmented system with $A$ sparse.

Uses an LDLt factorization of the quasi-definite augmented system.

Tulip.TLPLinearAlgebra.update_linear_solver! — Method

update_linear_solver!(ls, θ, regP, regD)

Update LDLt factorization of the augmented system.

Update diagonal scaling $\theta$, primal-dual regularizations, and re-compute the factorization. Throws a PosDefException if matrix is not quasi-definite.

Tulip.TLPLinearAlgebra.solve_augmented_system! — Method

solve_augmented_system!(dx, dy, ls, ξp, ξd)

Solve the augmented system, overwriting dx, dy with the result.

SparsePosDefLinearSolver

Tulip.TLPLinearAlgebra.SparsePosDefLinearSolver — Type

SparsePosDefLinearSolver{Tv}

Linear solver for the 2x2 augmented system

\[ [-(Θ^{-1} + Rp) A'] [dx] = [xi_d] [ A Rd] [dy] = [xi_p]\]

with $A$ sparse.

Uses a Cholesky factorization of the positive definite normal equations system

(A*(Θ^{-1} + Rp)^{-1}*A' + Rd)  dy = xi_p + A*(Θ^{-1} + Rp)^{-1}*xi_d
                                dx = (Θ^{-1} + Rp)^{-1} * (A' dy - xi_d)

Tulip.TLPLinearAlgebra.update_linear_solver! — Method

update_linear_solver!(ls, θ, regP, regD)

Compute normal equation system matrix, and update the factorization.

Tulip.TLPLinearAlgebra.solve_augmented_system! — Method

solve_augmented_system!(dx, dy, ls, ξp, ξd)

Solve the augmented system, overwriting dx, dy with the result.

LDLFLinearSolver

Tulip.TLPLinearAlgebra.LDLFLinearSolver — Type

LDLFLinearSolver{Tv}

Linear solver for the 2x2 augmented system with $A$ sparse.

Uses LDLFactorizations.jl to compute an LDLt factorization of the quasi-definite augmented system.

Tulip.TLPLinearAlgebra.update_linear_solver! — Method

update_linear_solver!(ls, θ, regP, regD)

Update LDLt factorization of the augmented system.

Update diagonal scaling $\theta$, primal-dual regularizations, and re-compute the factorization. Throws a PosDefException if matrix is not quasi-definite.

Tulip.TLPLinearAlgebra.solve_augmented_system! — Method

solve_augmented_system!(dx, dy, ls, ξp, ξd)

Solve the augmented system, overwriting dx, dy with the result.