Solving linear systems

The interior-point algorithm in Tulip requires the solution, at each iteration, of the following symmetric quasi-definite augmented system

\[\begin{bmatrix} -(\Theta^{-1} + R_{p}) & A^{T}\\ A & R_{d} \end{bmatrix} \begin{bmatrix} \Delta x\\ \Delta y \end{bmatrix} = \begin{bmatrix} \xi_d\\ \xi_p \end{bmatrix}\]

where

$\Delta x, \Delta y$ are primal and dual search directions,

$A$ is the problem's constraint matrix,

$\Theta$, $R_p$ and $R_d$ are positive diagonal matrices,

$\xi_p, \xi_d$ are right-hand side vectors.

The augmented system above can be reduced to the positive-definite normal equations system

\[\begin{array}{rl} \left( A (\Theta^{-1} + R_{p})^{-1} A^{T} + R_{d} \right) \Delta y & = \xi_{p} + A (Θ^{-1} + R_{p})^{-1} \xi_{d}\\ \Delta x &= (Θ^{-1} + R_{p})^{-1} (A^{T} \Delta y - \xi_{d}) \end{array}\]

When selected, this reduction is transparent to the interior-point algorithm.

To enable the use of fast external libraries and/or specialized routines, the resolution of linear systems is performed by a linear solver object. Linear solvers can be customized with the options below.

Linear solver options

System

Tulip.TLPLinearAlgebra.LinearSystem — Type

LinearSystem

Indicates which linear system is solved.

Tulip.TLPLinearAlgebra.DefaultSystem — Type

DefaultSystem

Choose linear system to be solved using default option.

Tulip.TLPLinearAlgebra.AugmentedSystem — Type

AugmentedSystem

Solve the augmented system, i.e., without reducing to normal equations.

Tulip.TLPLinearAlgebra.NormalEquations — Type

NormalEquations

Solve the normal equations system.

Backend

Tulip.TLPLinearAlgebra.LSBackend — Type

LSBackend

Backend used for solving linear systems.

Tulip.TLPLinearAlgebra.DefaultBackend — Type

DefaultBackend

Chose linear solver backend automatically.

For A::Matrix{Tv}, defaults to Lapack (i.e., dense solver).

For A::AbstractMatrix{Float64}, defaults to Cholmod

Otherwise, defaults to LDLFact

Tulip.TLPLinearAlgebra.Lapack — Type

Lapack <: LSBackend

Use LAPACK backend.

Options available:

Numerical precision: Tv<:Real

Float32 and Float64 will use LAPACK.

Other numerical types will use Julia's generic cholesky factorization.

NormalEquations only, with Cholesky factorization

Tulip.TLPLinearAlgebra.Cholmod — Type

Cholmod <: LSBackend

Use CHOLMOD backend.

Options available:

Float64 only

Augmented system with LDLᵀ factorization

Normal equations with Cholesky factorization

Tulip.TLPLinearAlgebra.LDLFact — Type

LDLFact <: LSBackend

Use LDLFactorizations backend.

Options available:

Any numerical type T

Augmented system with LDLᵀ factorization

Supported linear solvers

Here is a list of currently supported linear solvers:

Linear solver type	Tv	System	LSBackend	Method

DenseLinearSolver	Real	NormalEquations	Lapack	Cholesky

SparseIndefLinearSolver	Float64	AugmentedSystem	Cholmod	LDLᵀ

SparsePosDefLinearSolver	Float64	NormalEquations	Cholmod	Cholesky

LDLFLinearSolver	Real	AugmentedSystem	LDLFact	LDLᵀ

Default options

If no option is specified, then the linear solver is chosen as follows:

$A$ is dense: DenseLinearSolver

If $A$ sparse and Tv is Float64: SparseIndefLinearSolver

All other cases: LDLFLinearSolver