Overview

AbstractKKTSolver

This is the base type from which all implementations should derive.

Tulip.KKT.AbstractKKTSolver — Type

AbstractKKTSolver{T}

Abstract container for solving an augmented system

    [-(Θ⁻¹ + Rp)   Aᵀ] [dx] = [ξd]
    [   A          Rd] [dy]   [ξp]

where ξd and ξp are given right-hand side.

Custom linear solvers should inherit from the AbstractKKTSolver class, and extend the following two functions:

Tulip.KKT.update! — Function

update!(kkt, θinv, regP, regD)

Update internal data and factorization/pre-conditioner.

After this call, kkt can be used to solve the augmented system

    [-(Θ⁻¹ + Rp)   Aᵀ] [dx] = [ξd]
    [   A          Rd] [dy]   [ξp]

for given right-hand sides ξd and ξp.

Arguments

kkt::AbstractKKTSolver{T}: the KKT solver object

θinv::AbstractVector{T}: $θ⁻¹$

regP::AbstractVector{T}: primal regularizations

regD::AbstractVector{T}: dual regularizations

Tulip.KKT.solve! — Function

solve!(dx, dy, kkt, ξp, ξd)

Solve the symmetric quasi-definite augmented system

    [-(Θ⁻¹ + Rp)   Aᵀ] [dx] = [ξd]
    [   A          Rd] [dy]   [ξp]

and over-write dx, dy with the result.

Arguments

dx, dy: Vectors of unknowns, modified in-place

kkt: Linear solver for the augmented system

ξp, ξd: Right-hand-side vectors

Choosing between linear solvers

Tulip.KKT.SolverOptions — Type

SolverOptions

Used to pass options and instantiate KKT solvers.

Tulip.TLPLinearAlgebra.MatrixOptions — Type

MatrixOptions

Used to pass options for setting-up the matrix.

julia> model = Tulip.Model{Float64}()
julia> model.params.MatrixOptions = Tulip.TLA.MatrixOptions(SparseMatrixCSC)

Dense/LAPACK

Tulip.KKT.DenseSPD — Type

DenseSPD{T}

KKT solver for dense linear systems.

Uses a Cholesky factorization of the normal equations system. Not recommended for systems of size larger than a few thousands.

model = Tulip.Model{Float64}()
model.params.KKTOptions = Tulip.KKT.SolverOptions(Tulip.KKT.DenseSPD)

Info

Dispatches to BLAS/LAPACK in Float32 and Float64 arithmetic.

CHOLMOD

Tulip.KKT.CholmodSolver — Type

CholmodSolver

Uses CHOLMOD's factorization routines to solve the augmented system.

To use an LDLᵀ factorization of the augmented system (see CholmodSQD)

model.params.KKTOptions = Tulip.KKT.SolverOptions(CholmodSolver, normal_equations=false)

To use a Cholesky factorization of the normal equations system (see CholmodSPD)

model.params.KKTOptions = Tulip.KKT.SolverOptions(CholmodSolver, normal_equations=true)

Warning

CHOLMOD can only be used with Float64 arithmetic.

Tulip.KKT.CholmodSQD — Type

CholmodSQD

Linear solver for the 2x2 augmented system with $A$ sparse.

Uses an LDLᵀ factorization of the quasi-definite augmented system.

Tulip.KKT.CholmodSPD — Type

CholmodSPD

Linear solver for the 2x2 augmented system

    [-(Θ⁻¹ + Rp)   Aᵀ] [dx] = [xi_d]
    [   A          Rd] [dy]   [xi_p]

with $A$ sparse.

Uses a Cholesky factorization of the positive definite normal equations system

(A * ((Θ⁻¹ + Rp)⁻¹ * Aᵀ + Rd) dy = xi_p + A * (θ⁻¹ + Rp)⁻¹ * xi_d
                              dx = (Θ⁻¹ + Rp)⁻¹ * (Aᵀ * dy - xi_d)

LDLFactorizations

Tulip.KKT.LDLFactSQD — Type

LDLFactSQD{T}

Linear solver for the 2x2 augmented system with $A$ sparse.

Uses LDLFactorizations.jl to compute an LDLᵀ factorization of the quasi-definite augmented system.

model.params.KKTOptions = Tulip.KKT.SolverOptions(LDLFactSQD)

Krylov

Warning

Iterative methods are still an experimental feature. Some numerical and performance issues should be expected.

Tulip.KKT.KrylovSPD — Type

KrylovSPD{T, F, Tv, Ta}

Krylov-based Symmetric Positive-Definite (SPD) linear solver.

Applies a Krylov method to the normal equations systems, then recovers a solution to the augmented system. The selected Krylov method must therefore handle symmetric positive-definite systems. Suitable methods are CG or CR.

A KrylovSPD is selected as follows

model.params.KKTOptions = Tulip.KKT.SolverOptions(
    KrylovSPD, method=Krylov.cg,
    atol=1e-12, rtol=1e-12
)

The method argument is a function f::F whose signature must match

dy, _ = f(S, ξ; atol, rtol)

where S, ξ and dy are the normal equations system's left- and right-hand side, and solution vector, respectively. S may take the form of a matrix, or of a suitable linear operator.

Tulip.KKT.KrylovSID — Type

KrylovSID{T, F, Tv, Ta}

Krylov-based Symmetric InDefinite (SID) linear solver.

Applies a Krylov method to solve the augmented system, without exploiting its quasi-definiteness properties. The selected Krylov method must therefore handle symmetric indefinite systems. Suitable methods are MINRES or MINRES-QLP.

A KrylovSID is selected as follows

model.params.KKTOptions = Tulip.KKT.SolverOptions(
    KrylovSID, method=Krylov.minres,
    atol=1e-12, rtol=1e-12
)

The method argument is a function f::F whose signature must match

Δ, _ = f(K, ξ; atol, rtol)

where K, ξ and Δ are the augmented system's left- and right-hand side, and solution vector, respectively. K may take the form of a matrix, or of a suitable linear operator.

Tulip.KKT.KrylovSQD — Type

KrylovSQD{T, F, Tv, Ta}

Krylov-based Symmetric Quasi-Definite (SQD) linear solver.

Applies a Krylov method to solve the augmented system, taking advantage of its 2x2 block structure and quasi-definiteness. The selected Krylov method must therefore handle 2x2 symmetric quasi-definite systems. Suitable methods are TriCG and TriMR.

A KrylovSID is selected as follows

model.params.KKTOptions = Tulip.KKT.SolverOptions(
    KrylovSQD, method=Krylov.trimr,
    atol=1e-12, rtol=1e-12
)

The method argument is a function f::F whose signature must match

dy, dx, _ = f(A, ξp, ξd; M, N, atol, rtol)

where the augmented system is of the form

[ -N⁻¹ Aᵀ ] [dx] = [ξd]
[  A   M⁻¹] [dy] = [ξp]

i.e., $N = (Θ^{-1} + R_{p})^{-1}$ and $M = R_{d}^{-1}$.