AbstractLinearSolver

This is the base type from which all implementations should derive.

Tulip.TLPLinearAlgebra.AbstractLinearSolver — Type

AbstractLinearSolver{Tv}

Abstract container for solving an augmented system

    [-(Θ⁻¹ + Rp)   Aᵀ] [dx] = [ξd]
    [   A          Rd] [dy]   [ξp]

where ξd and ξp are given right-hand side.

Custom linear solvers should inherit from the AbstractLinearSolver class, and extend the following two functions:

Tulip.TLPLinearAlgebra.update_linear_solver! — Function

update_linear_solver!(ls, θ, regP, regD)

Update internal data, and re-compute factorization/pre-conditioner.

After this call, ls can be used to solve the augmented system

    [-(Θ⁻¹ + Rp)   Aᵀ] [dx] = [ξd]
    [   A          Rd] [dy]   [ξp]

for given right-hand sides ξd and ξp.

Tulip.TLPLinearAlgebra.solve_augmented_system! — Function

solve_augmented_system!(dx, dy, ls, ξp, ξd)

Solve the symmetric quasi-definite augmented system

    [-(Θ⁻¹ + Rp)   Aᵀ] [dx] = [ξd]
    [   A          Rd] [dy]   [ξp]

and over-write dx, dy with the result.

Arguments

dx, dy: Vectors of unknowns, modified in-place

ls: Linear solver for the augmented system

ξp, ξd: Right-hand-side vectors

DenseLinearSolver

Tulip.TLPLinearAlgebra.DenseLinearSolver — Type

DenseLinearSolver{Tv}

Linear solver for the 2x2 augmented system

    [-(Θ⁻¹ + Rp)   Aᵀ] [dx] = [xi_d]
    [   A          Rd] [dy]   [xi_p]

with $A$ dense.

The augmented system is automatically reduced to the normal equations system. BLAS/LAPACK functions are used whenever applicable.

Tulip.TLPLinearAlgebra.update_linear_solver! — Method

update_linear_solver!(ls::DenseLinearSolver{<:Real}, θ, regP, regD)

Compute normal equations system matrix and update Cholesky factorization.

Uses Julia's generic linear algebra.

Tulip.TLPLinearAlgebra.update_linear_solver! — Method

update_linear_solver!(ls::DenseLinearSolver{<:BlasReal}, θ, regP, regD)

Compute normal equations system matrix and update Cholesky factorization.

Uses BLAS and LAPACK routines.

Tulip.TLPLinearAlgebra.solve_augmented_system! — Method

solve_augmented_system!(dx, dy, ls, ξp, ξd)

Solve the augmented system, overwriting dx, dy with the result.

Uses two generic triangular solves for solving the normal equations system.

Tulip.TLPLinearAlgebra.solve_augmented_system! — Method

solve_augmented_system!(dx, dy, ls, ξp, ξd)

Solve the augmented system, overwriting dx, dy with the result.

Uses one LAPACK call for solving the normal equations system.

SparseIndefLinearSolver

Tulip.TLPLinearAlgebra.SparseIndefLinearSolver — Type

SparseIndefLinearSolver

Linear solver for the 2x2 augmented system with $A$ sparse.

Uses an LDLᵀ factorization of the quasi-definite augmented system.

Tulip.TLPLinearAlgebra.update_linear_solver! — Method

update_linear_solver!(ls, θ, regP, regD)

Update LDLᵀ factorization of the augmented system.

Update diagonal scaling $\theta$, primal-dual regularizations, and re-compute the factorization. Throws a PosDefException if matrix is not quasi-definite.

Tulip.TLPLinearAlgebra.solve_augmented_system! — Method

solve_augmented_system!(dx, dy, ls, ξp, ξd)

Solve the augmented system, overwriting dx, dy with the result.

SparsePosDefLinearSolver

Tulip.TLPLinearAlgebra.SparsePosDefLinearSolver — Type

SparsePosDefLinearSolver

Linear solver for the 2x2 augmented system

    [-(Θ⁻¹ + Rp)   Aᵀ] [dx] = [xi_d]
    [   A          Rd] [dy]   [xi_p]

with $A$ sparse.

Uses a Cholesky factorization of the positive definite normal equations system

(A * ((Θ⁻¹ + Rp)⁻¹ * Aᵀ + Rd) dy = xi_p + A * (θ⁻¹ + Rp)⁻¹ * xi_d
                              dx = (Θ⁻¹ + Rp)⁻¹ * (Aᵀ * dy - xi_d)

Tulip.TLPLinearAlgebra.update_linear_solver! — Method

update_linear_solver!(ls, θ, regP, regD)

Compute normal equation system matrix, and update the factorization.

Tulip.TLPLinearAlgebra.solve_augmented_system! — Method

solve_augmented_system!(dx, dy, ls, ξp, ξd)

Solve the augmented system, overwriting dx, dy with the result.

LDLFLinearSolver

Tulip.TLPLinearAlgebra.LDLFLinearSolver — Type

LDLFLinearSolver{Tv}

Linear solver for the 2x2 augmented system with $A$ sparse.

Uses LDLFactorizations.jl to compute an LDLᵀ factorization of the quasi-definite augmented system.

Tulip.TLPLinearAlgebra.update_linear_solver! — Method

update_linear_solver!(ls, θ, regP, regD)

Update LDLᵀ factorization of the augmented system.

Update diagonal scaling $\theta$, primal-dual regularizations, and re-compute the factorization. Throws a PosDefException if matrix is not quasi-definite.

Tulip.TLPLinearAlgebra.solve_augmented_system! — Method

solve_augmented_system!(dx, dy, ls, ξp, ξd)

Solve the augmented system, overwriting dx, dy with the result.