Tulip.get_attribute — Method

get_attribute(model::Model, ::BarrierIterations)

Query the BarrierIterations attribute from model

Tulip.get_attribute — Method

get_attribute(m::Model, ::ConstraintName, i::Int)

Query the name of constraint i in model m

Tulip.get_attribute — Method

get_attribute(model::Model, ::ModelName)

Query the ModelName attribute from model

Tulip.get_attribute — Method

get_attribute(model::Model, ::Status)

Query the Status attribute from model

Tulip.get_attribute — Method

get_attribute(m::Model{T}, ::VariableLowerBound, j::Int)

Query the lower bound of variable j in model m.

Tulip.get_attribute — Method

get_attribute(m::Model, ::VariableName, j::Int)

Query the name of variable j in model m

Tulip.get_attribute — Method

get_attribute(model::Model, ::DualObjectiveValue)

Query the DualObjectiveValue attribute from model

Tulip.get_attribute — Method

get_attribute(model::Model, ::ObjectiveValue)

Query the ObjectiveValue attribute from model

Tulip.get_parameter — Method

get_parameter(m::Model, pname::String)

Query the value of parameter pname in model m.

Tulip.load_problem! — Method

load_problem!(m::Model{T}, fname::String)

Read a model from file fname and load it into model m.

Only free MPS files are currently supported.

Tulip.set_attribute — Method

set_attribute(m::Model, ::ConstraintName, i::Int, name::String)

Set the name of constraint i in model m to name.

Tulip.set_attribute — Method

set_attribute(model::Model, ::ModelName, name::String)

Set the ModelName attribute in model

Tulip.set_attribute — Method

set_attribute(m::Model, ::VariableName, j::Int, name::String)

Set the name of variable j in model m to name.

Tulip.set_attribute — Method

set_attribute(m::Model{T}, ::ConstraintLowerBound, i::Int, lb::T)

Set the lower bound of constraint i in model m to lb.

Tulip.set_attribute — Method

set_attribute(m::Model{T}, ::ConstraintUpperBound, i::Int, ub::T)

Set the upper bound of constraint i in model m to ub.

Tulip.set_attribute — Method

set_attribute(m::Model{T}, ::VariableLowerBound, j::Int, lb::T)

Set the lower bound of variable j in model m to lb.

Tulip.set_attribute — Method

set_attribute(m::Model{T}, ::VariableUpperBound, j::Int, ub::T)

Set the upper bound of variable j in model m to ub.

Tulip.set_parameter — Method

set_parameter(m::Model, pname::String, val)

Set the value of parameter pname in model m to val